update project & fix bug

+++ admin/src/Pages/Categegory/editCategory.jsx
@@ -59,18 +59,24 @@ const EditCategory = () => {
     }
 
     const removeImg = (image, index) => {
-        var imageArr = [];
-        imageArr = previews;
-        deleteImages(`/api/category/deteleImage?img=${image}`).then((res) => {
-            imageArr.splice(index, 1);
+        context?.showConfirmDelete(
+            "Xóa ảnh danh mục?",
+            "Bạn có chắc chắn muốn xóa ảnh danh mục này?",
+            () => {
+                var imageArr = [];
+                imageArr = previews;
+                deleteImages(`/api/category/deteleImage?img=${image}`).then((res) => {
+                    imageArr.splice(index, 1);
 
-            setPreviews([]);
-            setTimeout(() => {
-                setPreviews(imageArr);
-                formFields.images = previewsArr
-            }, 100);
+                    setPreviews([]);
+                    setTimeout(() => {
+                        setPreviews(imageArr);
+                        formFields.images = imageArr
+                    }, 100);
 
-        })
+                })
+            }
+        )
     }
 
 
@@ -80,13 +86,13 @@ const EditCategory = () => {
         setIsLoading(true);
 
         if (formFields.name === "") {
-            context.alertBox("error", "Please enter category name");
+            context.alertBox("error", "Vui lòng nhập tên danh mục");
             setIsLoading(false);
             return false
         }
 
         if (previews?.length === 0) {
-            context.alertBox("error", "Please select category image");
+            context.alertBox("error", "Vui lòng chọn ảnh danh mục");
             setIsLoading(false);
             return false
         }
@@ -117,7 +123,7 @@ const EditCategory = () => {
 
                     <br />
 
-                    <h3 className='text-[14px] font-[500] mb-2 text-black'> Categegory Image</h3>
+                    <h3 className='text-[14px] font-[500] mb-2 text-black'>Ảnh danh mục</h3>
   
                     <div className="grid grid-cols-2 md:grid-cols-7 gap-4">
                         {
@@ -152,7 +158,7 @@ const EditCategory = () => {
                                 :
                                 <>
                                     <FaCloudUploadAlt className='text-[25px] text-white' />
-                                    Publish and View
+                                    Lưu và xem
                                 </>
                         }
                     </Button>
